--- app/Http/Controllers/PasswordController.php
@@ -1,35 +1,0 @@
-<?php
-
-namespace App\Http\Controllers\Auth;
-
-use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Hash;
-use Illuminate\Validation\Rules\Password;
-
-class PasswordController extends Controller
-{
-    /**
-     * Update the user's password when forced to change it
-     */
-    public function forceUpdate(Request $request)
-    {
-        $request->validate([
-            'current_password' => ['required', function ($attribute, $value, $fail) {
-                if (!Hash::check($value, auth()->user()->password)) {
-                    $fail('The current password is incorrect.');
-                }
-            }],
-            'password' => ['required', 'confirmed', Password::min(8)],
-        ]);
-
-        // Update password
-        auth()->user()->update([
-            'password' => Hash::make($request->password),
-            'must_change_password' => false, // Actualizado a inglés
-        ]);
-
-        return redirect()->route('dashboard')
-            ->with('success', 'Your password has been changed successfully.');
-    }
-}
